Move DeleteTable into ActionsPane
