+++ mathesar_ui/src/sections/table-view/TableView.svelte
@@ -14,7 +14,7 @@
   } from '@mathesar/stores/tableData';
   import ActionsPane from './actions-pane/ActionsPane.svelte';
   import DisplayOptions from './display-options/DisplayOptions.svelte';
-  import DeleteTable from './DeleteTable.svelte';
+  import DeleteTable from './actions-pane/DeleteTable.svelte';
   import Header from './Header.svelte';
   import Body from './Body.svelte';
   import StatusPane from './status-pane/StatusPane.svelte';
