+++ mathesar_ui/src/sections/table-view/actions-pane/DeleteTable.svelte
